New: moved header into layout
