+++ client/src/layout/Gradient/gradient.module.css
@@ -3,6 +3,10 @@
   background-repeat: no-repeat;
   background-size: cover;
 
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
   min-height: 100vh;
   width: 100vw;
 }
