+++ client/src/layout/WithNavbar/types.tsx
@@ -1,0 +1,5 @@
+import { PropsWithChildren } from "react";
+
+export interface ILayoutWithNavbar extends PropsWithChildren {
+  headerTitle?: string;
+}
