+++ client/src/pages/notebook/goals/goals.module.css
@@ -1,12 +1,3 @@
-.header {
-  grid-column: 1 / -1;
-  grid-row: 1 / 2;
-
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-}
-
 .layout {
   display: grid;
   grid-template-columns: 1fr;
@@ -18,15 +9,13 @@
   width: 100%;
 }
 
-.header {
+.actions {
   grid-column: 1 / -1;
   grid-row: 1 / 2;
-}
 
-.actions {
   display: flex;
-  justify-content: space-between;
-  gap: 4rem;
+  justify-content: flex-end;
+  gap: 3rem;
 }
 
 .content {
@@ -40,19 +29,17 @@
 }
 
 .filter {
-  width: 16rem;
+  width: 100%;
+}
+
+.addButton {
+  min-width: 12rem;
 }
 
 @media only screen and (min-width: 600px) {
-  .header {
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-between;
-  }
-
-  .actions {
-    justify-content: unset;
-  }
+  .filter {
+  width: 16rem;
+}
 }
 
 @media only screen and (min-width: 900px) {
