+++ client/src/pages/notebook/notebook.module.css
@@ -13,7 +13,7 @@
 @media screen and (min-width: 900px) {
   .layout {
     grid-template-columns: repeat(2, 1fr);
-    grid-template-rows: min-content repeat(2, 1fr);
+    grid-template-rows: repeat(2, 1fr);
     grid-gap: 3rem;
   }
