+++ client/src/pages/notebook/notes/Suggestions/suggestions.module.css
@@ -1,6 +1,6 @@
 .ideas {
   grid-column: 1 / -1;
-  grid-row: 2 / 3;
+  grid-row: 1 / 2;
 
   display: flex;
   flex-direction: column;
