+++ client/src/pages/notebook/payments/payments.module.css
@@ -1,17 +1,3 @@
-.grid {
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: min-content 1fr;
-
-  height: 100%;
-  width: 100%;
-}
-
-.header {
-  grid-column: 1 / -1;
-  grid-row: 1 / 2;
-}
-
 .container {
   grid-row: 2 / 3;
