+++ client/src/pages/notebook/workhours/workhours.module.css
@@ -10,7 +10,7 @@
 
 .tabsWrapper {
   grid-column: 1 / 6;
-  grid-row: 2;
+  grid-row: 1 / 2;
 }
 
 .addContainer {
@@ -36,36 +36,31 @@
 @media only screen and (min-width: 900px) {
   .layout {
     grid-template-columns: repeat(8, 1fr);
-    grid-template-rows: max-content max-content repeat(5, 1fr) max-content;
+    grid-template-rows: auto repeat(5, 1fr) max-content;
     column-gap: 1rem;
     row-gap: 2rem;
   }
 
-  .header {
-    grid-column: 1 / 6;
-    grid-row: 1 / 2;
-  }
-
   .tracker {
     grid-column: 1 / 6;
-    grid-row: 3 / 8;
+    grid-row: 2 / 7;
   }
 
   .addContainer {
     grid-column: 6 / 9;
-    grid-row: 4 / -1;
+    grid-row: 3 / -1;
   }
 
   .timer {
     grid-column: 6 / 9;
-    grid-row: 1 / 4;
+    grid-row: 1 / 3;
 
     display: block;
   }
 
   .reminder {
     grid-column: 1 / 6;
-    grid-row: 8 / 8;
+    grid-row: 7 / 8;
   }
 
   .addForm {
