--- client/src/pages/resources/emergencyNumbers/emergencyNumbers.module.css
@@ -1,16 +1,0 @@
-.layout {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: flex-start;
-  gap: 2rem;
-
-  height: 100%;
-  width: 100%;
-}
-
-.layout ul {
-  margin: 0;
-  margin-left: 1rem;
-  padding: 0;
-}
