+++ client/src/pages/resources/index.tsx
@@ -1,10 +1,8 @@
-import { Resources as ResourcesList, Text } from "components";
-
+import { Resources as ResourcesList } from "components";
 import styles from "./resources.module.css";
 
 export const Resources = (): JSX.Element => (
   <div className={ styles.layout }>
-    <Text as="h2" size="xl" weight="bold">Resources</Text>
     <ResourcesList />
   </div>
 );
