+++ client/src/pages/resources/resources.module.css
@@ -6,15 +6,3 @@
   height: 100%;
   width: 100%;
 }
-
-.header {
-  grid-column: 1 / -1;
-  grid-row: 1 / 2;
-}
-
-.info {
-  grid-column: 1 / -1;
-  grid-row: 2 / 3;
-}
-
-@media only screen and (min-width: 900px) {}
