+++ client/src/pages/resources/tax/index.tsx
@@ -10,7 +10,6 @@ export const Tax = (): JSX.Element => {
 
   const renderHeader = useMemo(() => headerRows.cells.map( ( cell, index ) => (
     <Fragment key={ index }>
-      <Text as="h3" size="xl" className={ styles.header } weight="bold">{ cell.title }</Text>
       <p className={ styles.info } dangerouslySetInnerHTML={ createMarkup( cell.text ) } />
     </Fragment>
   ) ), [headerRows]);
